feat: provide `sept11 nist assoc-dir-with-tape` cmd
I'm beginning a new repository for management of the tasks I create in my personal Todoist setup.
The previous one was getting difficult to manage, with each task type in a different script file. In
this new setup, I'm going for a single interface that will use different commands to create the
different types of tasks I need.

The first command, `sept11 nist assoc-dir-with-tape`, creates a new task in the '9/11 Archiving'
project for associating one of the directories in the releases with a tape in NIST's database.

I'll carry over commands from the previous repository as and when I need them.