feat: provide `food new-item` cmd
Food items can be saved in a SQLite database. These items are going to be used as the basis for
forming meals, unless they are a snack that can be eaten on their own. I initially thought of having
the whole meal as a food item; that's why a `MealType` enum is included in the commit, but not used.
I think it will be quicker over time if I compose meals from items, because existing items can be
re-used.