feat: provide `food new-meal` cmd
A meal consists of one or more food items, each of which has a quantity. We can compose meals by
re-using existing food items, which should make things easier over time. The quantity is to
facilitate, e.g., a meal that would consist of two sandwiches, which has four slices of bread. The
slice of bread would only need to be added as one food item.