feat: provide `food plan` cmd
A simple interface enables the user to select meals or snacks to add to a meal plan for a selected
date. The command would have to run multiple times for working with different dates. This should
hopefully work well as a first attempt, and it can be extended as need be.