feat: provide `add-shopping-item` cmd
Select food items from the database to add to the shopping list project. The meal plan can be
reviewed on a daily basis to determine which items need to be purchased, and these can go on the
list. The list can be consulted on the mobile app to try and help me not forget the things I need to
buy.