chore: reorganise command menu
In anticipation of adding commands for listing food items and meals, the menu is reworked to
introduce more subcommands. Placeholders for the new commands are provided.

The text on the menus were also updated with a more accurate description of what the commands are
doing.

BREAKING CHANGE: previous commands have now been deleted.