feat: provide `food items rm` command
Remove a food item from the database.

It can only be removed it if hasn't already been used as part of a meal. We would need to provide a
command for deleting the meal first. In my current use case, I made a mess of an item before adding
it to a meal, so for now I don't need to remove any meals.